+++ ocrd/ocrd/workspace_bagger.py
@@ -6,7 +6,7 @@ from tempfile import mkdtemp
 import re
 import tempfile
 import sys
-from bagit import Bag, make_manifests  # pylint: disable=no-name-in-module
+from bagit import Bag, make_manifests, _load_tag_file  # pylint: disable=no-name-in-module
 
 from ocrd_utils import (
     pushd_popd,
@@ -58,7 +58,7 @@ class WorkspaceBagger():
         else:
             log.info(msg)
 
-    def _bag_mets_files(self, workspace, bagdir, ocrd_manifestation_depth, ocrd_mets, processes):
+    def _bag_mets_files(self, workspace, bagdir, ocrd_mets, processes):
         mets = workspace.mets
         changed_urls = {}
 
@@ -67,13 +67,10 @@ class WorkspaceBagger():
         with pushd_popd(workspace.directory):
             # URLs of the files before changing
             for f in mets.find_files():
-                log.info("Resolving %s (%s)", f.url, ocrd_manifestation_depth)
+                log.info("Resolving %s", f.url)
                 if is_local_filename(f.url):
                     # nothing to do then
                     pass
-                elif ocrd_manifestation_depth != 'full':
-                    self._log_or_raise("Not fetching non-local files, skipping %s" % f.url)
-                    continue
                 elif not f.url.startswith('http'):
                     self._log_or_raise("Not an http URL: %s" % f.url)
                     continue
@@ -96,7 +93,7 @@ class WorkspaceBagger():
         # Walk through bagged workspace and fix the PAGE
         # Page/@imageFilename and
         # AlternativeImage/@filename
-        bag_workspace = Workspace(self.resolver, directory=join(bagdir, 'data'))
+        bag_workspace = Workspace(self.resolver, directory=join(bagdir, 'data'), mets_basename=ocrd_mets)
         with pushd_popd(bag_workspace.directory):
             for page_file in bag_workspace.mets.find_files(mimetype=MIMETYPE_PAGE):
                 pcgts = page_from_file(page_file)
@@ -118,7 +115,7 @@ class WorkspaceBagger():
             log.info("New vs. old: %s" % changed_urls)
         return total_bytes, total_files
 
-    def _set_bag_info(self, bag, total_bytes, total_files, ocrd_identifier, ocrd_manifestation_depth, ocrd_base_version_checksum):
+    def _set_bag_info(self, bag, total_bytes, total_files, ocrd_identifier, ocrd_base_version_checksum, ocrd_mets='mets.xml'):
         bag.info['BagIt-Profile-Identifier'] = OCRD_BAGIT_PROFILE_URL
         bag.info['Bag-Software-Agent'] = 'ocrd/core %s (bagit.py %s, bagit_profile %s) [cmdline: "%s"]' % (
             VERSION, # TODO
@@ -127,18 +124,18 @@ class WorkspaceBagger():
             ' '.join(sys.argv))
 
         bag.info['Ocrd-Identifier'] = ocrd_identifier
-        bag.info['Ocrd-Manifestation-Depth'] = ocrd_manifestation_depth
         if ocrd_base_version_checksum:
             bag.info['Ocrd-Base-Version-Checksum'] = ocrd_base_version_checksum
         bag.info['Bagging-Date'] = str(datetime.now())
         bag.info['Payload-Oxum'] = '%s.%s' % (total_bytes, total_files)
+        if ocrd_mets != 'mets.xml':
+            bag.info['Ocrd-Mets'] = ocrd_mets
 
     def bag(self,
             workspace,
             ocrd_identifier,
             dest=None,
             ocrd_mets='mets.xml',
-            ocrd_manifestation_depth='full',
             ocrd_base_version_checksum=None,
             processes=1,
             skip_zip=False,
@@ -155,15 +152,12 @@ class WorkspaceBagger():
             ord_identifier (string): Ocrd-Identifier in bag-info.txt
             dest (string): Path of the generated OCRD-ZIP.
             ord_mets (string): Ocrd-Mets in bag-info.txt
-            ord_manifestation_depth (string): Ocrd-Manifestation-Depth in bag-info.txt
             ord_base_version_checksum (string): Ocrd-Base-Version-Checksum in bag-info.txt
             processes (integer): Number of parallel processes checksumming
             skip_zip (boolean): Whether to leave directory unzipped
             in_place (boolean): Whether to **replace** the workspace with its BagIt variant
             tag_files (list<string>): Path names of additional tag files to be bagged at the root of the bag
         """
-        if ocrd_manifestation_depth not in ('full', 'partial'):
-            raise Exception("manifestation_depth must be 'full' or 'partial'")
         if in_place and (dest is not None):
             raise Exception("Setting 'dest' and 'in_place' is a contradiction")
         if in_place and not skip_zip:
@@ -194,11 +188,11 @@ class WorkspaceBagger():
             f.write(BAGIT_TXT.encode('utf-8'))
 
         # create manifests
-        total_bytes, total_files = self._bag_mets_files(workspace, bagdir, ocrd_manifestation_depth, ocrd_mets, processes)
+        total_bytes, total_files = self._bag_mets_files(workspace, bagdir, ocrd_mets, processes)
 
         # create bag-info.txt
         bag = Bag(bagdir)
-        self._set_bag_info(bag, total_bytes, total_files, ocrd_identifier, ocrd_manifestation_depth, ocrd_base_version_checksum)
+        self._set_bag_info(bag, total_bytes, total_files, ocrd_identifier, ocrd_base_version_checksum, ocrd_mets=ocrd_mets)
 
         for tag_file in tag_files:
             copyfile(tag_file, join(bagdir, basename(tag_file)))
@@ -239,6 +233,7 @@ class WorkspaceBagger():
 
         bagdir = mkdtemp(prefix=TMP_BAGIT_PREFIX)
         unzip_file_to_dir(src, bagdir)
+        bag_info = _load_tag_file(join(bagdir, "bag-info.txt"))
 
         datadir = join(bagdir, 'data')
         for root, _, files in walk(datadir):
@@ -259,7 +254,8 @@ class WorkspaceBagger():
         rmtree(bagdir)
 
         # Create workspace
-        workspace = Workspace(self.resolver, directory=dest)
+        mets_basename = bag_info.get("Ocrd-Mets", "mets.xml")
+        workspace = Workspace(self.resolver, directory=dest, mets_basename=mets_basename)
 
         # TODO validate workspace