+++ ocrd_utils/ocrd_utils/constants.py
@@ -43,6 +43,7 @@ EXT_TO_MIME = {
     '.pnm': 'image/x-portable-anymap',
     '.pbm': 'image/x-portable-bitmap',
     '.txt': 'text/plain',
+    '.xsl': 'text/xsl',
 }
 
 MIME_TO_EXT = {
@@ -57,10 +58,15 @@ MIME_TO_EXT = {
     'application/pdf': '.pdf',
     'application/postscript': '.ps',
     'application/oxps': '.xps',
+    'application/x-hdf': '.h5',
+    'application/x-hdf;subtype=bag': '.h5',
+    'application/vnd.pytorch': '.pth',
     'image/x-portable-pixmap': '.ppm',
     'image/x-portable-anymap': '.pnm',
     'image/x-portable-bitmap': '.pbm',
     'text/plain': '.txt',
+    'text/xsl': '.xsl',
+    'text/xml': '.xml',
 }
 
 #