+++ ocrd_utils/ocrd_utils/os.py
@@ -3,17 +3,22 @@ Operating system functions.
 """
 __all__ = [
     'abspath',
+    'directory_size',
     'is_file_in_directory',
+    'get_ocrd_tool_json',
+    'get_moduledir',
     'get_processor_resource_types',
     'pushd_popd',
     'unzip_file_to_dir',
     'atomic_write',
 ]
 
-from tempfile import TemporaryDirectory
+from tempfile import TemporaryDirectory, gettempdir
+from functools import lru_cache
 import contextlib
 from distutils.spawn import find_executable as which
 from json import loads
+from json.decoder import JSONDecodeError
 from os import getcwd, chdir, stat, chmod, umask, environ
 from pathlib import Path
 from os.path import exists, abspath as abspath_, join, isdir
@@ -23,6 +28,7 @@ from subprocess import run, PIPE
 from atomicwrites import atomic_write as atomic_write_, AtomicWriter
 
 from .constants import XDG_DATA_HOME
+from .logging import getLogger
 
 def abspath(url):
     """
@@ -42,16 +48,16 @@ def pushd_popd(newcwd=None, tempdir=False):
         oldcwd = getcwd()
     except FileNotFoundError:
         # This happens when a directory is deleted before the context is exited
-        oldcwd = '/tmp'
+        oldcwd = gettempdir()
     try:
         if tempdir:
             with TemporaryDirectory() as tempcwd:
                 chdir(tempcwd)
-                yield tempcwd
+                yield Path(tempcwd).resolve()
         else:
             if newcwd:
                 chdir(newcwd)
-            yield newcwd
+            yield Path(newcwd).resolve()
     finally:
         chdir(oldcwd)
 
@@ -63,68 +69,117 @@ def unzip_file_to_dir(path_to_zip, output_directory):
     z.extractall(output_directory)
     z.close()
 
-def list_resource_candidates(executable, fname, cwd=getcwd()):
+@lru_cache()
+def get_ocrd_tool_json(executable):
+    """
+    Get the ``ocrd-tool`` description of ``executable``.
+    """
+    executable_name = Path(executable).name
+    try:
+        ocrd_tool = loads(run([executable, '--dump-json'], stdout=PIPE).stdout)
+    except (JSONDecodeError, OSError) as e:
+        getLogger('ocrd_utils.get_ocrd_tool_json').error(f'{executable} --dump-json produced invalid JSON: {e}')
+        ocrd_tool = {}
+    if 'resource_locations' not in ocrd_tool:
+        ocrd_tool['resource_locations'] = ['data', 'cwd', 'system', 'module']
+    return ocrd_tool
+
+@lru_cache()
+def get_moduledir(executable):
+    moduledir = None
+    try:
+        moduledir = run([executable, '--dump-module-dir'], encoding='utf-8', stdout=PIPE).stdout.rstrip('\n')
+    except (JSONDecodeError, OSError) as e:
+        getLogger('ocrd_utils.get_moduledir').error(f'{executable} --dump-module-dir failed: {e}')
+    return moduledir
+
+def list_resource_candidates(executable, fname, cwd=getcwd(), moduled=None, xdg_data_home=None):
     """
     Generate candidates for processor resources according to
-    https://ocr-d.de/en/spec/ocrd_tool#file-parameters (except python-bundled)
+    https://ocr-d.de/en/spec/ocrd_tool#file-parameters
     """
     candidates = []
     candidates.append(join(cwd, fname))
+    xdg_data_home = XDG_DATA_HOME if not xdg_data_home else xdg_data_home
     processor_path_var = '%s_PATH' % executable.replace('-', '_').upper()
     if processor_path_var in environ:
         candidates += [join(x, fname) for x in environ[processor_path_var].split(':')]
-    candidates.append(join(XDG_DATA_HOME, 'ocrd-resources', executable, fname))
+    candidates.append(join(xdg_data_home, 'ocrd-resources', executable, fname))
     candidates.append(join('/usr/local/share/ocrd-resources', executable, fname))
+    if moduled:
+        candidates.append(join(moduled, fname))
     return candidates
 
-def list_all_resources(executable):
+def list_all_resources(executable, moduled=None, xdg_data_home=None):
     """
     List all processor resources in the filesystem according to
-    https://ocr-d.de/en/spec/ocrd_tool#file-parameters (except python-bundled)
+    https://ocr-d.de/en/spec/ocrd_tool#file-parameters
     """
     candidates = []
+    try:
+        resource_locations = get_ocrd_tool_json(executable)['resource_locations']
+    except FileNotFoundError:
+        # processor we're looking for ressource_locations of is not installed.
+        # Assume the default
+        resource_locations = ['data', 'cwd', 'system', 'module']
+    xdg_data_home = XDG_DATA_HOME if not xdg_data_home else xdg_data_home
     # XXX cwd would list too many false positives
-    # cwd_candidate = join(getcwd(), 'ocrd-resources', executable)
-    # if Path(cwd_candidate).exists():
-    #     candidates.append(cwd_candidate)
+    # if 'cwd' in resource_locations:
+    #     cwd_candidate = join(getcwd(), 'ocrd-resources', executable)
+    #     if Path(cwd_candidate).exists():
+    #         candidates.append(cwd_candidate)
     processor_path_var = '%s_PATH' % executable.replace('-', '_').upper()
     if processor_path_var in environ:
         for processor_path in environ[processor_path_var].split(':'):
             if Path(processor_path).is_dir():
                 candidates += Path(processor_path).iterdir()
-    datadir = Path(XDG_DATA_HOME, 'ocrd-resources', executable)
-    if datadir.is_dir():
-        candidates += datadir.iterdir()
-    systemdir = Path('/usr/local/share/ocrd-resources', executable)
-    if systemdir.is_dir():
-        candidates += systemdir.iterdir()
+    if 'data' in resource_locations:
+        datadir = Path(xdg_data_home, 'ocrd-resources', executable)
+        if datadir.is_dir():
+            candidates += datadir.iterdir()
+    if 'system' in resource_locations:
+        systemdir = Path('/usr/local/share/ocrd-resources', executable)
+        if systemdir.is_dir():
+            candidates += systemdir.iterdir()
+    if 'module' in resource_locations and moduled:
+        # recurse fully
+        for resource in itertree(Path(moduled)):
+            if resource.is_dir():
+                continue
+            if any(resource.match(pattern) for pattern in
+                   # Python distributions do not distinguish between
+                   # code and data; `is_resource()` only singles out
+                   # files over directories; but we want data files only
+                   # todo: more code and cache exclusion patterns!
+                   ['*.py', '*.py[cod]', '*~', 'ocrd-tool.json', 
+                    'environment.pickle', 'resource_list.yml', 'lib.bash']):
+                continue
+            candidates.append(resource)
     # recurse once
     for parent in candidates:
-        if parent.is_dir():
+        if parent.is_dir() and parent.name != '.git':
             candidates += parent.iterdir()
-    return [str(x) for x in candidates]
+    return sorted([str(x) for x in candidates])
 
 def get_processor_resource_types(executable, ocrd_tool=None):
     """
-    Determine whether a processor has resource parameters that represent
-    directories (``has_dirs``), files (``has_files``) or neither.
+    Determine what type of resource parameters a processor needs.
 
-    Returns a pair ``(has_dir, has_files)``
+    Return a list of MIME types (with the special value `*/*` to
+    designate that arbitrary files or directories are allowed).
     """
     if not ocrd_tool:
         # if the processor in question is not installed, assume both files and directories
         if not which(executable):
-            return (True, True)
-        result = run([executable, '--dump-json'], stdout=PIPE, check=True, universal_newlines=True)
-        ocrd_tool = loads(result.stdout)
-    if not next((True for p in ocrd_tool['parameters'].values() if 'content-type' in p), False):
+            return ['*/*']
+        ocrd_tool = get_ocrd_tool_json(executable)
+    if not next((True for p in ocrd_tool.get('parameters', {}).values() if 'content-type' in p), False):
         # None of the parameters for this processor are resources (or not
         # the resource parametrs are not properly declared, so output both
         # directories and files
-        return (True, True)
-    has_dirs = next((True for p in ocrd_tool['parameters'].values() if p.get('content-type', None) == 'text/directory'), False)
-    has_files = next((True for p in ocrd_tool['parameters'].values() if 'content-type' in p and p['content-type'] != 'text/directory'), False)
-    return (has_dirs, has_files)
+        return ['*/*']
+    return [p['content-type'] for p in ocrd_tool['parameters'].values()
+            if 'content-type' in p]
 
 # ht @pabs3
 # https://github.com/untitaker/python-atomicwrites/issues/42
@@ -155,3 +210,21 @@ def is_file_in_directory(directory, file):
     directory = Path(directory)
     file = Path(file)
     return list(file.parts)[:len(directory.parts)] == list(directory.parts)
+
+def itertree(path):
+    """
+    Generate a list of paths by recursively enumerating ``path``
+    """
+    if not isinstance(path, Path):
+        path = Path(path)
+    if path.is_dir():
+        for subpath in path.iterdir():
+            yield from itertree(subpath)
+    yield path
+
+def directory_size(path):
+    """
+    Calculcates size of all files in directory ``path``
+    """
+    path = Path(path)
+    return sum(f.stat().st_size for f in path.glob('**/*') if f.is_file())