+++ tests/cli/test_workspace.py
@@ -131,6 +131,7 @@ class TestCli(TestCase):
         content = 'x'
         mimetype = 'image/tiff'
         with TemporaryDirectory() as tempdir:
+            tempdir = str(Path(tempdir).resolve())
             content_file = join(tempdir, 'testfile')
             with open(content_file, 'w') as f:
                 f.write(content)
@@ -236,6 +237,7 @@ class TestCli(TestCase):
         file_grp = 'TEST_GROUP'
         mimetype = 'image/tiff'
         with TemporaryDirectory() as tempdir:
+            tempdir = str(Path(tempdir).resolve())
             content_file = join(tempdir, 'test.tif')
             ws = self.resolver.workspace_from_nothing(directory=tempdir)
             ws.save_mets()
@@ -287,7 +289,7 @@ class TestCli(TestCase):
             self.assertEqual(result.exit_code, 0)
 
             ws2 = self.resolver.workspace_from_url(join(tempdir, 'ws', 'mets.xml'))
-            self.assertEqual(len(ws2.mets.find_all_files()), 7)
+            self.assertEqual(len(ws2.mets.find_all_files()), 29)
 
     def test_clone_into_nonexisting_dir(self):
         """
@@ -308,6 +310,7 @@ class TestCli(TestCase):
         Test removal of filegrp
         """
         with TemporaryDirectory() as tempdir:
+            tempdir = str(Path(tempdir).resolve())
             wsdir = join(tempdir, 'ws')
             copytree(assets.path_to('SBB0000F29300010000/data'), wsdir)
             file_group = 'OCR-D-GT-PAGE'