+++ tests/test_task_sequence.py
@@ -10,11 +10,14 @@ from tests.data.wf_testcase import (
     PARAM_JSON,
 )
 
-from ocrd_utils import pushd_popd, MIMETYPE_PAGE
+from ocrd_utils import pushd_popd, MIMETYPE_PAGE, get_ocrd_tool_json
 from ocrd.resolver import Resolver
 from ocrd.task_sequence import run_tasks, validate_tasks, ProcessorTask
 
 class TestOcrdWfStep(TestCase):
+
+    def tearDown(self):
+        get_ocrd_tool_json.cache_clear()
 
     def test_parse_no_in(self):
         task = ProcessorTask.parse('sample-processor')