+++ CHANGELOG.md
@@ -4,7 +4,38 @@ Change Log
 Versioned according to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
-p
+
+Fixed:
+
+  * `ocrd resmgr download *` working again, #904, #908, #909
+  * Resource manager respects `moduledir` correctly, #904
+
+Added:
+
+  * Processors have a `--dump-module-dir` to print their implementation-specific module directory to STDOUT, #904
+
+## [2.38.0] - 2022-08-14
+
+Fixed:
+
+  * `ocrd zip`: Properly respect `Ocrd-Mets`, #899
+  * `ocrd workspace merge`: missing arguments, #896
+  * `ocrd resmgr download`: Support dynamic discovery, #901
+
+Added:
+
+  * Processors support profiling with `--profile` and `--profile-file`, #878, bertsky/core#4
+
+Removed:
+
+  * `ocrd zip`: remove support for obsolete `Ocrd-Manifestation-Depth`, #902, OCR-D/spec#182
+
+## [2.37.0] - 2022-08-03
+
+Added:
+
+  * `ocrd resmgr`: Resources of processors can be described in the `ocrd-tool.json`, #800
+
 ## [2.36.0] - 2022-07-18
 
 Fixed:
@@ -1504,6 +1535,8 @@ Fixed
 Initial Release
 
 <!-- link-labels -->
+[2.38.0]: ../../compare/v2.38.0..v2.37.0
+[2.37.0]: ../../compare/v2.37.0..v2.36.0
 [2.36.0]: ../../compare/v2.36.0..v2.35.0
 [2.35.0]: ../../compare/v2.35.0..v2.34.0
 [2.34.0]: ../../compare/v2.34.0..v2.33.0